Moves bean status finding logic into a separate class for use with QueueControllerService.
